+++ script/seed-lineitems.js
@@ -1,0 +1,24 @@
+const lineitems = [
+  {
+    orderId: 1,
+    productId: 1,
+    purchasedPrice: 10,
+    quantity: 1
+  },
+
+  {
+    orderId: 2,
+    productId: 2,
+    purchasedPrice: 200,
+    quantity: 2
+  },
+
+  {
+    orderId: 3,
+    productId: 3,
+    purchasedPrice: 300,
+    quantity: 3
+  }
+]
+
+module.exports = lineitems
