+++ script/seed-products.js
@@ -1,0 +1,29 @@
+const products = [
+  {
+    name: 'book worm duck',
+    description:
+      'We all love reading a good book, so why shouldnt your rubber duck enjoy one too? Please note, this duck does not squeak',
+    price: 100,
+    imageUrl:
+      'https://cdn.shopify.com/s/files/1/1184/9194/products/literary-rubber-duck-1322-p_600x.jpeg?v=1457991497'
+  },
+
+  {
+    name: 'happy duck',
+    description:
+      'Life is even more fun with this cheerful rubber duck! This duck becomes everyones playmate',
+    price: 200,
+    imageUrl:
+      'https://www.internet-toys.com/producten/large/free_and_easy_badeend_26_x_20_x_22_cm_geel_268798_20190103155226.jpg'
+  },
+
+  {
+    name: 'sad duck',
+    description: 'Lets make this sad duck happy buying it',
+    price: 300,
+    imageUrl:
+      'https://images-na.ssl-images-amazon.com/images/I/51UNZiX8D-L._AC_SX425_.jpg'
+  }
+]
+
+module.exports = products
