+++ script/seed.js
@@ -5,85 +5,9 @@ const db = require('../server/db')
 const {User, Product, Order, LineItem} = require('../server/db/models')
 
 const users = require('./seed-users')
-
-const products = [
-  {
-    name: 'book worm duck',
-    description:
-      'We all love reading a good book, so why shouldnt your rubber duck enjoy one too? Please note, this duck does not squeak',
-    price: 100,
-    imageUrl:
-      'https://cdn.shopify.com/s/files/1/1184/9194/products/literary-rubber-duck-1322-p_600x.jpeg?v=1457991497'
-  },
-
-  {
-    name: 'happy duck',
-    description:
-      'Life is even more fun with this cheerful rubber duck! This duck becomes everyones playmate',
-    price: 200,
-    imageUrl:
-      'https://www.internet-toys.com/producten/large/free_and_easy_badeend_26_x_20_x_22_cm_geel_268798_20190103155226.jpg'
-  },
-
-  {
-    name: 'sad duck',
-    description: 'Lets make this sad duck happy buying it',
-    price: 300,
-    imageUrl:
-      'https://images-na.ssl-images-amazon.com/images/I/51UNZiX8D-L._AC_SX425_.jpg'
-  }
-]
-
-const orders = [
-  {
-    totalQuantity: 1,
-    totalCost: 100,
-    shipping_Date: new Date(2020, 8, 19),
-    delivery_Date: new Date(2020, 9, 29),
-    shipping_address: 'Hanover street, NewYork',
-    status: 'Placed'
-  },
-  {
-    totalQuantity: 2,
-    totalCost: 400,
-    shipping_Date: new Date(2020, 8, 19),
-    delivery_Date: new Date(2020, 9, 29),
-    shipping_address: 'Hanover street, NewYork',
-    status: 'Cart'
-  },
-
-  {
-    totalQuantity: 3,
-    totalCost: 900,
-    shipping_Date: new Date(2020, 8, 19),
-    delivery_Date: new Date(2020, 9, 29),
-    shipping_address: 'Hanover street, NewYork',
-    status: 'Placed'
-  }
-]
-
-const lineitems = [
-  {
-    orderId: 1,
-    productId: 1,
-    purchasedPrice: 10,
-    quantity: 1
-  },
-
-  {
-    orderId: 2,
-    productId: 2,
-    purchasedPrice: 200,
-    quantity: 2
-  },
-
-  {
-    orderId: 3,
-    productId: 3,
-    purchasedPrice: 300,
-    quantity: 3
-  }
-]
+const products = require('./seed-products')
+const orders = require('./seed-orders')
+const lineitems = require('./seed-lineitems')
 
 const seed = async () => {
   try {
